% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclone_energy.R
\name{cyclone_energy}
\alias{cyclone_energy}
\title{Compute Accumulated Cyclone Energy for given storm(s)}
\usage{
cyclone_energy(stormid)
}
\arguments{
\item{stormid}{A character or a character vector of storm IDs in the HURDAT
dataset. The ID should be in the format similar to "AL182012" or a vector
of multiple IDs.}
}
\value{
A data frame with storm ID(s) and corresponding ACE value(s).
}
\description{
This function calculates the Accumulated Cyclone Energy (ACE) of a given set
of storm(s) by their storm ID in the HURDAT dataset.
}
\examples{
cyclone_energy(c("AL182012"))
cyclone_energy(c("AL182012", "AL011851"))
}
