% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landfall_pos.R
\name{landfall_pos}
\alias{landfall_pos}
\title{Determine if a storm made landfall in the US}
\usage{
landfall_pos(storm.id, date, time)
}
\arguments{
\item{storm.id}{a character representing the ID of the storm in the
HURDAT dataset.}

\item{date}{a character representing the date of the storm in the
format "YYYYMMDD".}

\item{time}{a character representing the time of the storm in the
format "HHMM".}
}
\value{
a logical value indicating whether the storm made landfall within the
border of the United States at the given date and time.
}
\description{
This function determines whether a storm with a given ID made landfall within
the border of the United States at a given date and time.
}
\examples{
landfall_pos("AL182012", "20121029", "1200")
}
