% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_next_q5(1).R, R/position_size.R
\name{pos_next}
\alias{pos_next}
\title{Given the point position of a storm, get the coordinate of the position of
the storm based on the moving distance and moving angle.}
\usage{
pos_next(pos0, dist, ang)

pos_next(pos0, dist, ang)
}
\arguments{
\item{pos0, }{the coordinate of the original point position of a storm}

\item{dist, }{moving distnce}

\item{ang, }{angle of the moving storm}

\item{arg1}{Description of argument 1.}

\item{arg2}{Description of argument 2.}
}
\value{
pos, a vector contains x-coordinate and y-coordinate of the next
position of the storm

Description of the return value.
}
\description{
Detailed description of the function.
}
\examples{
pos_next <- pos_next(c(-50,30), 10, 45)
my_function(arg1 = 1, arg2 = "abc")
}
