% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_size.R
\name{position_size}
\alias{position_size}
\title{Map the position and size of a given storm}
\usage{
position_size(stormid, date, time)
}
\arguments{
\item{stormid}{a character representing the hurricane ID in the HURDAT dataset.
The ID should be in this format similar to "AL182012".}

\item{date}{a character representing the date in the "YYYYMMDD" format.}

\item{time}{a character representing the time in the "HHMM" format}
}
\value{
a ggplot object showing the storm's position and size on a map
}
\description{
This R function takes the storm ID, date, and time as input and generates a
plot of the storm position and its corresponding size of 34 knots, 50 knots,
and 64 knots on a map. This can be useful for visualizing the intensity and
location of a tropical storm.
}
\examples{
position_size("AL182012", "20121029", "2100")

}
