% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_30min_incre.R
\name{storm_30min_incre}
\alias{storm_30min_incre}
\title{Generate 30-minute interval tracks for the input storm ID(s)}
\usage{
storm_30min_incre(storm.id)
}
\arguments{
\item{storm.id}{A character or a character vector of valid storm ID(s).}
}
\value{
A dataframe containing the 30-minute interval tracks for each input
storm ID(s).
}
\description{
This function generates 30-minute interval tracks for the input storm ID(s),
containing corresponding information of id, date, time, numeric.latitude,
numeric.longitude.
}
\examples{
# Generate 30-minute interval tracks for the storm "AL011851"
storm_ids <- "AL011851"
storm_30min_incre(storm_ids)
# Generate 30-minute interval tracks for the storm "AL011851" and "AL021851"
storm_ids <- c("AL011851", "AL021851")
storm_30min_incre(storm_ids)
}
