% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_map.R
\name{storm_map}
\alias{storm_map}
\title{Brief description of the function}
\usage{
storm_map(storm.ids)
}
\arguments{
\item{arg1}{Description of argument 1.}

\item{arg2}{Description of argument 2.}
}
\value{
Description of the return value.
}
\description{
Detailed description of the function.
}
\examples{
my_function(arg1 = 1, arg2 = "abc")
}
