% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_30min_incre.R
\name{stormz_30min_incre}
\alias{stormz_30min_incre}
\title{Brief description of the function}
\usage{
stormz_30min_incre(storm.id)
}
\arguments{
\item{arg1}{Description of argument 1.}

\item{arg2}{Description of argument 2.}
}
\value{
Description of the return value.
}
\description{
Detailed description of the function.
}
\examples{
my_function(arg1 = 1, arg2 = "abc")
}
