% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclone_energy.R
\name{cyclone_energy}
\alias{cyclone_energy}
\title{Brief description of the function}
\usage{
cyclone_energy(storm.ids)
}
\arguments{
\item{arg1}{Description of argument 1.}

\item{arg2}{Description of argument 2.}
}
\value{
Description of the return value.
}
\description{
Detailed description of the function.
}
\examples{
my_function(arg1 = 1, arg2 = "abc")
}
