% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_landfall.R
\name{storm_landfall}
\alias{storm_landfall}
\title{Identify if storm(s) landfell within the border of the United States}
\usage{
storm_landfall(storm.ids)
}
\arguments{
\item{storm.ids}{A character or a character vector of storm IDs}
}
\value{
A logical vector indicating whether each storm landfell within the
border of the United States
}
\description{
This function takes in a storm ID or a vector of storm IDs and identifies
if each storm landfell within the border of the United States.
}
\examples{
# Identify if the storm "AL052005" landfell within the border of the United States
storm_ids <- "AL052005"
storm_landfall(storm_ids)
# Identify which storms landfell within the border of the United States
storm_ids <- c("AL052005", "AL062005", "AL072005")
storm_landfall(storm_ids)
}
\keyword{hurricane}
\keyword{landfall}
\keyword{storm}
