% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_map.R
\name{storm_map}
\alias{storm_map}
\title{Generate a map of storm track(s) given storm ID(s)}
\usage{
storm_map(storm.ids, title = "")
}
\arguments{
\item{storm.ids}{A character or a character vector of valid storm ID(s) to
include in the map.}

\item{title}{An optional character vector for the title of the plot. Default
is an empty string.}
}
\value{
A ggplot2 object representing the storm tracks map.
}
\description{
This function generates a map of storm tracks given a storm ID or a vector of
storm ID(s).
}
\examples{
# Generate a map of storm tracks for hurricanes with ID "AL052021"
storm_map("AL052021")
# Generate a map of storm tracks for hurricanes with ids "AL052021" and "AL072021"
storm_map(c("AL052021", "AL072021"))
}
